Updated openstack/openstack Project: openstack/trove  5b53ac27b2e43a90956eeb5a068841a566559aac
Add os_admin to the ignored_dbs for PostgreSQL

Recently we added an os_admin user to PostgreSQL for Trove use.
PostgreSQL requires a database for each user, so this now appears in the
database-list command. It needs to be added to the 'ignore_dbs' config
option so that this doesn't happen.

The tests don't capture this since they use the cfg setting to make sure
that no 'unwanted' db names are returned - a bit of a chicken/egg
problem.

Change-Id: Ia900826d9f2d775466dabde433e9ac1349148112
Closes-Bug: #1621997