Updated openstack/openstack Project: openstack/trove-integration  26087fd58c4d9e39cdf12171266cb21c9f18e1bd
Speed up creation of flavors

When creating the flavors necessary for redstack int-tests
'nova flavor-list' is executed for each flavor.  This adds
over a minute to the kick-start process, even if all the
flavors already exists.

This change caches the list of flavors and passes it into
the add_flavor routine to avoid the overhead of multiple
flavor-list calls.

Change-Id: Ib70efd4eeb6f6808360d9baaac14edd5a6e84934