Updated openstack/openstack Project: openstack/nova  ec2e03f908fd0fedf6eb49ef1fb484e9b53b062e
Compute: ensure that InvalidDiskFormat is handled correctly

In the event that a hypervisor does not support a specific disk format
it will raise an exception InvalidDiskFormat. In this circumstance
the instance was being stuck in 'SCHEDULING' state.

When a exception like this is encountered we will treat it similarly
to the ImageUnacceptable and InvalidDiskInfo execptions.

Change-Id: I61ed418e3884c773df87718d13fef6f69010342b
Closes-bug: #1621564