Updated openstack/openstack Project: openstack-infra/tripleo-ci  e6f536516f989f1ce904f21f6aa7e8d4874f05e8
Limit Heat engine workers to 4 in rh1

If we have more than 4 Heat engine workers in rh1 it can generate
so much traffic that the other services have trouble keeping up,
which causes all kinds of trouble.  4 seems to be a pretty good
sweet spot where Heat has plenty of capacity to create stacks but
doesn't DoS the other services.

Note that this is already in the running config on rh1.  This is
just a patch to update the env file so it will be persistent.

Change-Id: I2c5d33cf2307349ea231ad1ba07170b250a84cef