Updated openstack/openstack Project: openstack/heat  1f2cb45040b14463cf191df35b9a662aebdcce2c
Convergence: basic framework for cancelling workers

Implements mechanism to cancel existing workers (in_progress resources).
The stack-cancel-update request lands in one of the engines, and if
there are any workers in that engine which are working for the stack,
they are cancelled first and then other engines are requested to cancel
the workers.

Change-Id: I464c4fdb760247d436473af49448f7797dc0130d