Updated openstack/openstack Project: openstack/neutron  4bfc1d58975cd75d15b812eb67197f2c16bcfc4f
Add sanity pings to restart tests to ensure connection

This adds sanity pings to the functional agent restart tests
to make sure there is actual connectivity between the two namespaces
before starting the restart exercise. This helps distinguish
whether a failure is because of a setup issue or because of
the restart process.

Change-Id: Id64ce49597b7e62e065f840a1372a930ce558e38
Partial-Bug: #1621789