Updated openstack/openstack Project: openstack/neutron  daac2701138169d20a41356cb26ff5e28d684c06
Emit router payload only during failures

When everything works as expected, no-one hardly pays any attention
to this log trace, which accounts for an incredible amount of log data.
This change proposes to emit the router payload only during failures
(when debugging info is needed the most), and furthermore it relocates
it to the L3 agent log files, where it is more pertinent.

Partial-bug: #1620864

Change-Id: I64281b963ba52c0a100a6194b7cafc5e9b1a8e74