Updated openstack/openstack Project: openstack/neutron  952be63acd4e65ff0b1aad8baa4af8599f0a5390
Reduce log level for ryu in OVS agent log

In a tempest run, Ryu is logging about ~30K debug traces and it puts the
OVS agent's log amongst the biggest log files in a single node OpenStack
deployment in the gate. This patch sets Ryu and its components to log at
INFO level and above to reduce the number of traces emitted.

This patch, alongside [1,2], brings down the size of the biggest Neutron
log files, server and OVS agent respectively. More can surely be done,
however callbacks and ryu are the most obvious ones, and following up
with further cuts may be dealt with on a ad-hoc basis.

Closes-bug: #1620864

[1] https://review.openstack.org/#/c/366424/
[2] https://review.openstack.org/#/c/366478/
Change-Id: I40b7c77601788ae2e7428c7a0206ca2a807d10dc