Updated openstack/openstack Project: openstack/nova  ff6b192aede1f1362f2c5252a50d9b5ffe201e3b
neutron: don't trace on port not found when unbinding ports

There is a race in the gate when deleting an instance and
deallocating the network at the same time that preexisting
ports attached to the instance are being deleted. So when
nova goes to unbind the port it's already gone and we log
an exception trace in the n-cpu logs. We shouldn't actually
care if the port isn't found when unbinding it from an instance,
so this change handles that case and logs it at debug rather
than an exception trace.

Change-Id: Ia95c626cefcb1e099e11d3bf5a651ad5d5f9406f
Closes-Bug: #1621883