Updated openstack/openstack Project: openstack/nova  0d83e08da4cf6ea1c5f8216841f7561c83798352
Add sync_power_state_pool_size option

The sync_power_state_pool_size option allows to set
the number of greenthreads available for use to sync power states.

It can be used to reduce the number of concurrent requests
made to the hypervisor or system with real instance power states
for performance reasons.

Closes-bug: #1621987
Change-Id: I9cf900314d71c44ec51eb190c102ddc60e13a767