Updated openstack/openstack Project: openstack-infra/project-config  0416f479f106ea887db0535fcee78d8415dde31c
Disable ovh-bhs1 until image uploads have been fixed

For a while now, image uploads to ovh-bhs1 have been problematic. As a
result, the images we are using are currently out of date.
Additionally, because we cannot upload images properly, it is starting
to affect uploads in other clouds. We only have 16 threads for
uploads.

Change-Id: I8894b03e5af1e4355e03da681ecfd2d0c6c510f3
Signed-off-by: Paul Belanger <pabelanger@redhat.com>