Updated openstack/openstack Project: openstack/neutron  cb99ce5c4465dbe8b63a5d2444ed6a6099b472a7
LinuxBridge: Pass host into get_devices_details_list

Pass the host into get_devices_details_list on the linux bridge
agent so the debug logs on the server side don't show "host None".
This is mainly just for cosmetics and consistency with the OVS agent
since the only thing the host is really used for on the server side
is special treatment of DVR ports, which does not currently apply
to linux bridge.

Change-Id: I700fa26982bdb087cf7ea4b3eb69aec2f2e099c8
Closes-Bug: #1622566