Updated openstack/openstack Project: openstack/manila  0a2fcb86bd9f128a5ad12e9218e34f0c724cd2e8
NetApp cDOT driver autosupport broken

The autosupport functionality of the NetApp cDOT driver is not working
with Python 2.7.12, the default in Xenial. Root cause is a deep copy
of the controller connection context, which includes some SSL context
stuff that doesn't play well with copy.deepcopy. So we can do a shallow
copy instead.

Change-Id: Ia2adc4ce27834e384e6d994fcb012ebf1d97c85c
Closes-Bug: #1621260