Updated openstack/openstack Project: openstack/oslo.config  5b700c7b21ca0135ee26b21f0ea449282a3185d1
Fix repr for tuples in choices

Using the repr on an object which had used a tuple for the value in
choices, would cause an error.

  TypeError: not all arguments converted during string formatting

Make code consistent for __repr__() with use of 'format'. And make code
consistent with use of 'if self.choices is not None'

Add additional tuple related unit tests

Change-Id: I76bb23b45a14348456572ad3af5bd6500a58fc96
Closes-Bug: #1621673