Updated openstack/openstack Project: openstack/tripleo-heat-templates  6dbcec15d35312149a5ab79c5524d292bdbcef10
De-bracket vncproxy_host in compute profile

This is done in the vncproxy profile, but for some reason is not in
the compute one.  It causes hiera to explode when the brackets are
left, so we need to do the bracket stripping here too.

Also switches both places to just use the host_nobrackets version
of the endpoint instead of stripping them with str_replace.

Change-Id: I7ccd84b575fd652f6412fdb1869c31c79a7bf53b
Closes-Bug: 1618623