Updated openstack/openstack Project: openstack/neutron  d6d53a795908059d2706cfdda5d7fc13f32cea92
Skip can_be_trunked validation for subports

A subport is not subject to the enforcement of this rule, which
applies only to parent ports. On the other hand, we should prevent
accidental use of a device port in a trunk because of unknown
side effects.

This patch tweaks the subport validation logic and ensures that
when a port is being used as a subport in a trunk, the following
rules are applied:

 * association rules: a record does not exist in neither tables,
   trunks and subports.
 * business logic rule: the port is not already assigned to another
   device.

Partially-implements: blueprint vlan-aware-vms

Change-Id: I7eeae5a0bff647d3faaa8996b37cea32ce654a62