Updated openstack/openstack Project: openstack/neutron  adee80ea90499ca02a92cd8310d8cc7f64f8e8d1
Set device_owner for trunk subports

Altough this is not currently required by any of the business
logic, it is handy to have this set to help users quickly
identify ports used in trunks. Furthermore, by adding this now
we avoid the potential headache of having to deal with a
post-upgrade situation where existing ports miss this important
piece of information.

Partially-implements: blueprint vlan-aware-vms

Change-Id: Icadefd86177643d153d0a6f615d551f6193b1f70