Updated openstack/openstack Project: openstack/heat  945c6e23b3cc1c5cb4fc20d542bff4215fc1c145
Revert "Improve validation for some functions"

This reverts commit 36bf170f347f927fd7f20b5868dc686041081862.

The construction and validation phases of intrinsic functions are
deliberately separated, to:

- Reduce the amount of unnecessary work done
- Reduce the risk of changes to validation breaking the loading of existing
  stacks from the database
- Ensure that the most critical errors are reported to the user first

This patch reverts a commit that sought to combine the two phases. There
was no linked bug and no explanation other than that this 'improved'
things.

It does not, however, restore the check for the Removed function during
properties translation (and the associated test), because I believe its
original existence was based on a misconception. There is no way in
practice for a hot.functions.Removed object to appear in the properties a
translation rule is operating on.

Change-Id: I59d02e540771baf4fe078081cd0993a3251eb9a6