Updated openstack/openstack Project: openstack/neutron  dedb632ba50e3df395fed33dcc70a2529b43524e
Defer setting 'ha'/'distributed' flags in L3 code

Both DVR and the HA code were setting the 'ha' and 'distributed'
flags in the API body before it was being sent into the core L3
code. This meant that it could not distinguish between
user-requested flags and config-defaults, which is important for
flavor validation.

This patch just adjusts it so they aren't set until after the core
create method is called.

Long term these will be refactored to live in their corresponding
driver anyway and will not need to be responsible for setting these
flags to get them stored in the DB.

Closes-Bug: #1621430
Change-Id: I9945920d5540653cf5b86e8f1a2ba7b073595921