Updated openstack/openstack Project: openstack/trove  1ef945d6fa8012486b625a98d3baab8db2b4ca12
Implement Instance Upgrade

Implments Instance Upgrade functionality to support upgrading the
image of a Trove datastore instance from datastore_version to a
newer datastore_version of the same datastore.

This functionality builds on the Nova rebuild API to upgrade
the image of an instance with minimal downtime.

Includes datastore implementation of the upgrade functionality
for the Mysql based datastores.

Change-Id: Ie6e48d78ac07df52f686f359ca7fdadaae6ad064
Implements: blueprint image-upgrade
Depends-On: I6ec2ebb78019c014f87ba5d8cbfd284686c64f30