Updated openstack/openstack Project: openstack-infra/storyboard  4926b471e2ccf0f4adb6e92e9f26769c942ae621
Stop allowing user's data to be found via private stories

Currently, private fields such as email addresses are not filtered
out of the user data sent in the permissions list of private stories.

This commit fixes that by correctly calling _filter_private_fields on
the list of users to be returned.

Change-Id: I4fa84c6787f758455288287af9bcd5c93159b172