Updated openstack/openstack Project: openstack/neutron  01dc8ab61017369ab77eea9bdeb46e673f02c3ce
objects: don't call DeclarativeObject's __init__ twice for rbac objects

Due to the way we applied metaclasses to rbac aware objects, the
__init__ was executed twice: once when the DeclarativeObject metaclass
was applied to NeutronDbObject, and once when rbac metaclass was
applied, when we were calling type() in rbac metaclass' __new__.

To avoid that behaviour, we can introduce a separate base 'class' for
rbac aware objects that would be a result of with_metaclass call using
the rbac metaclass, and NeutronDbOjbect as the actual base. Then we use
the base class in place of NeutronDbObject base class for all objects
that are RBAC aware.

This avoids the duplicate call issue.

Change-Id: Ia12f56f7d40c38e679edc21e13871e27ad1820ee
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db