Updated openstack/openstack Project: openstack/heat  d1a43447ecd4827d2b7e93a6d795d33ebc7dd8a2
Ensure cached conditions match stack

Ensure that the caching we do of conditions in TemplateCommon always
applies to the stack that has been passed in as an argument.

This was not the case previously, and in fact it was required not to be the
case because in the convergence path the new parameter values were not
copied into the stack at the start of an update. This patch rectifies that.

Change-Id: I942b53570222bf4e0d1bf78e584792d445c14090