Updated openstack/openstack Project: openstack/neutron  e07b78c7359bf6b64291a2e78fcf86530e81216b
devstack: the last fix of is_kernel_supported_for_ovs25 condition

Previously we didn't consider kernels higher than 4. This patch returns
True right away if the verions is lesser than 4. If it is 4, then it must
be lesser or equal to 3 to return True. All other versions are not
supported.

Change-Id: I3e9fa088d7cb9cfecbe7670c84a051e15be2a3a9