Updated openstack/openstack Project: openstack/cinder  632b7961e293d83c29fe8f7b58b7b4d6d0d6cfd1
Add strict Boolean checking for storage pools

There is no strict boolean checking for the parameter
"detail" of API /scheduler-states/get_pools, so that
any invalid boolean value can be specified.
This patch adds a strict checking for it to prevent
invalid value, and adds a test for this change as well.

Change-Id: Ic24850e5a0e206548de81529179cd182d9eabb7f
Partial-Bug: #1594261