Updated openstack/openstack Project: openstack/governance  e2d60a26bc68bd21649079e64bab1e837d4e5ec4
Add tag for tempest plugins

This commit adds a new deliverable type tag to describe tempest
plugins. For right now this is just for standalone plugin repos.
However the vast majority of tempest plugins are packaged in the
project repos. This is actually an anti-pattern and makes consuming
it more difficult. [1] But, cleaning that up is a separate matter. In
the interim it might be worth adding the tag to those repos with
included plugins as well. But that is a separate discussion to have
and will happen in, a follow on change.

[1] http://docs.openstack.org/developer/tempest/plugin.html#standalone-plugin-vs-in-repo-plugin

Change-Id: Ia713945e88f9f207f9a260eb83b9e595629c4bc3