Updated openstack/openstack Project: openstack/nova  3c6b71d311d748dc379ed18c043d59098c734d2a
Update BuildRequest if instance currently being scheduled

There are some cases where the instance is still waiting to be scheduled and
consequently there is still a BuildRequest object where we would also like to
update the instance attributes, like the display name.

In that case, we need to update the nested instance field from the BuildRequest
instead of trying to update the instance object itself, which could be located
in a child cell.

Change-Id: I97cd6dbae10ca1a29f646b0b25bc08edd699edca
Closes-Bug: #1620248