Updated openstack/openstack Project: openstack/nova  11ff0ded4e6fe34ec7c0a152ff142e836956dda9
More conservative inventory updates

This makes the compute node check for the current inventory,
compare it to the one it would send to the placement service, and
only make the PUT call if updates are needed. This results in lower
write traffic to the DB, fewer generation conflicts, etc.

It also aggressively tries to refresh its view of the remote generation,
which is now changing on every allocation as well, to avoid needing to
retry inventory updates later.

Change-Id: Id846f184c569341ae96e1696cb95e7f901e27f66