Updated openstack/openstack Project: openstack-infra/zuul  50408bc9559ff6b4dd9c0c34421ac1587a4f7165
Ansible launcher: add AFS publisher

This is an extension to JJB that works only in zuul-launcher, not
Jenkins.  It allows copying the results of a build into afs.
It actually isn't really AFS specific at all, other than it
checks that the destination path is under /afs.  Otherwise, it
behaves as a local copy on the launcher itself.

It also contains the logic needed to publish OpenStack's
documentation builds, which can appear as subdirectories of other
builds.

Change-Id: Icda75266219d2d7167e80aaad8e290443cfdbadc