Updated openstack/openstack Project: openstack/keystone  c2fd1f6af54f4518dc29e71d0a9c15e80f608984
Fix links on configure_federation documentation

When the API reference was moved and the old pages cleaned up, a lot of
dead links were created. This patch fixes them for the documentation on
"Configuring Keystone for Federation".

Moreover, a lot of the link text was nondescriptive, which makes the
documentation inaccessible for screen readers (see the W3C
guideline[1]). This patch cleans that up as well if the link URL
needed to be updated anyway.

[1] https://www.w3.org/TR/WCAG20-TECHS/H30.html

Change-Id: I58803276d9b06bad0252da2494c81a46c951916f