Updated openstack/openstack Project: openstack/ceilometer  b9ef78748215dd535075ea9b40c7849a0f2155eb
Remove store_events options

The current option is misnamed, as it does not enable any storing but
just the processing of events from the notification agent to the
collector.

This means that even if you set event_dispatchers=panko and forget to
set store_events=true, nothing will happen.

This patch enable the event processing as soon as something is
configured in the pipeline.

Change-Id: I5a906684f6371b0548ac08cacc13aa238f780f78